Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Comments is misleading here.. deployment of DL feeder should be via blueprint as discussed under DCAEGEN2-1849. 
> 
> Other standalone deployment option may be included (possibly as helm chart) but for integration will likely need a blueprint using k8s plugin (which will create k8s service/deployment yaml).

This is not part of DCAEGEN2-1849. It is a nice-to-have and easy way for development and testing of DMaaP integration, to be used by developers. Before this, we run DataLake against standalone Kafka; with this approach, we now can run it against DMaaP, without the 'formal' DL deployment.

We are working on DCAEGEN2-1849, and this is the 1st related change: 
https://gerrit.onap.org/r/c/dcaegen2/services/+/99970

Patch-set: 1
Reviewer: Gerrit User 2349 <2349@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0