Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (6 comments)
> 
> bbs-event-processor/version.properties should be updated to reflect new version

These were the original files in BBS-ep (under 'dpo' directory).

├── blueprints
│   ├── bbs-event-processor-input.yaml
│   └── k8s-bbs-event-processor.yaml-template
├── data-formats
│   ├── aai-interaction.json
│   └── pnf-reregistration.json
└── spec
    └── bbs-event-processor-spec.json

Due to the introduction of the new repo for blueprints, I thought we would have to move complete 'dpo' directory into that. So, if I understand well, only blueprint-related files should be transferred and data-formats and spec should remain as before. Is that correct?

Piotr will comment for the keystores

Patch-set: 1
Label: Code-Review=0