Update patch set 1
Patch Set 1: Code-Review+1

With file type in the file name, it tells the reader what kind of script this file is (in this case it is a UNIX shell script), which is nice to know without looking into the file content.
"script names do not require an extension reflecting their implementation language" -- reflecting their implementation language is a nice thing to have. If it does not break anything, I suggest we keep the file extension.

Patch-set: 1
Reviewer: Gerrit User 2349 <2349@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1