Update patch set 3
Patch Set 3:

> Patch Set 1: Code-Review-1
> 
> (2 comments)
> 
> version bump required (as current - 1.1.0 is released already)

Sorry forgot to set this as WIP. I will be be abandoning this change. Looks like there is no need for any changes for the build error(SOCKET_TIMEOUT). I had raised a ticket to resolve this build error. They had manually downloaded the package and restarted the job and so it was successful. I think any commit in datalake-handler repo will require this manual action as you can see the build has failed again. If there is any other alternative to solve this build error it would be great. Not sure if this is due to temporary service downtime.

Patch-set: 3