Update patch set 1
Patch Set 1: Code-Review-1

It's hard to verify that the code is doing the right thing when the Jira ticket has no description for what it is supposed to do.

What is it supposed to do when it succeeds?
What is the correct response when it succeeds?
What is the correct response when a name is not found?
What is the correct response when an error occurs during the processing?

Patch-set: 1
Reviewer: Gerrit User 105 <105@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1