Update patch set 2
Patch Set 2: Code-Review-1

There's no description in the Jira ticket. What is the API supposed to do? What are the values supposed to contain? When would there be multiple values in the array? When should errors be returned? When should success be returned? What should happen if the measurement_group_name on the URI different from what is in the body?

Patch-set: 2
Reviewer: Gerrit User 105 <105@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1