Update patch set 7
Patch Set 7:

Regarding the copyright errors:

You need to use a consistent Copyright notice for YOUR edits in this push. Instead there is a mishmash of Copyright notices, some from Huawei and some from Huawei Canada.

For example, if you are with Huawei Canada, then ALL of the modified files need a line for 2022 Huawei Canada.

Patch-set: 7